Add "mbox" module
Verifies mail in a single UNIX mailbox file. Presentation is controled with
four variables: "prefix", "label", "label_no_mail", and "postfix". Note,
"prefix" and "postfix" are concatinated with "label" (or "label_no_mail") with
no additional spaces, so sufficient whitespace must be provided. By default, all
these are empty except for "label", which is set to filename (last component of
path to) mailbox.