New "mail" module
Checks for mail in mailboxes specified in mailpath. Supports maildir and mbox
mailboxes. If mailpath is not specified, system mailbox (/var/mail/username) is
checked.

Presentation of check results is controled with two variables: "format" and
"format_no_mail". Variable format has following syntax: "prefix{body}postfix";
"prefix" is printed once, then "body" is printed for each mailbox, and then
postfix is printed once. Alternatively, "format" may specify only "body" part,
with no braces; in such case title of the mailchecker will be used as prefix.
Keywords "%mailbox" and "%messages" are expanded in "body"; keyword "%total" -
in "prefix" and "postfix". Nothing is expanded in "format_no_mail".
The mbox implementation in this module expects "mboxo" type of mboxes, so it may
report wrong numbers for other types of mboxes.  In particular, it does not
check "Content-Length" header.
The maildir implementation supports Courier-style ("maildir++") nested maildirs.
Effectively it would search for directories containing subdirectory "new" and
count regular files there. It would detect maildirs creation and deletion in
flight.
Variable "ignore" may be set to names of mailboxes that should not be checked
for mail. The number of messages in ignored mailboxes will not affect total
message count as well.
The width of report may be limited by setting variable "width" to amount of
bytes to output. In such case the report will be trimmed, and a trim indicator
will be placed in the end of the string. Trim indicator may be altered by
setting variable "replacement".