Allow to print IPv6 addresses in ethernet and wireless modules
This adds the new format option "%ip6" which will be replaced by the
first IPv6 found for the corresponding interface.

In `print_eth_info` and `print_wireless_info` functions, the first call
to `get_ip_addr` is now used only to detect if the wanted interface is up
and has an IP. Other calls to `get_ip_addr` will be done while we parse
the format_up string in order to get and print the actual addresses.