Add new module to display short strings from file.
This new module allows i3status to display short character sequences
generated from an alternate script.  For example, an email monitoring
script can generate a sequence of character for each inbox that has
unread mail pending, storing the sequence in a file shared with the
i3status config.

That script can update the status file on a schedule independent on
the i3status update interval.  The static-text modules caches the
contents of the file, rereading it whenever it gets updated or
truncated.  So if the status file is updated infrequently, the only
cost to i3status is an additional stat(2) system call.