Disable GLOB_TILDE if it isn't available
GLOB_TILDE is a GNU extension that is not supported by all libcs; for
example, the glob() function of the musl libc doesn't support it.

GLOB_TILDE is currently used in two places in the i3status code, in
process_runs.c and print_cpu_temperature.c. However, the ability to
expand ~ isn't essential in either case; it is merely present as a
convenience for the user. Hence, we can just disable it if the libc
lacks support -- having to type a few extra characters in a config file
is still much better than not being able to build i3status at all.