travis: update dist, add custom matrix
New matrix for 5 jobs in total.

Updates our checks to bionic & xenial. Trusty could also be
included but it is a bit outdated at this point.

I use bionic to pull a more recent clang-format and use a dedicated job
just to check formatting. An alternative would be to conditionally run
the formatting script only on bionic.

I don't think that there is a better way to expand a `dist` x `compiler`
matrix (except using docker).