pulseaudio: Retry until ready (w/ timeout)
This is a minor issue but immediately after pulse_initialize() is called
for the first time, the condition
`if (!context_ready || default_sink_idx == DEFAULT_SINK_INDEX)`
in description_pulseaudio() fails.

If a pulse device is explicitly specified, the first volume that
i3status prints gets assigned to zero. Otherwise, with the default
device, the ALSA fallback is used, unnecessarily initializing other
resources.

This patch implements a small wait that is run only once during
i3status' lifetime.

Fixes a small memleak that for some reason happens with the ALSA
fall-back if pulseaudio is used.