Add another battery status called 'idle'
On some systems, a battery's charging regulator may not charge a battery
even though it is not full. This might be the case because it was
configured to stop charging at a capacity threshold, or e.g. because
environmental conditions do not allow for charging the battery.

This commit adds this status (called 'idle') and adds support for
detecting the status on Linux.