Revert "Use the PulseAudio port name rather than the sink name (#419)"
Changing the port name is not as straighforward as changing the device
description.
Changing the device description is easy and supported in multiple
modern media backends such as pipewire-media-session and Wireplumber.

Also, the device description is unique most of the time in default
deployments. The port name is often the same across different devices
such as "Digital Output (S/PDIF)" or "Speakers". This leads to confusion
when changing the output device via hotkeys and checking the active
device via i3status in uncustomized i3status configurations.

Furthermore, in pavucontrol, when assigning output/input devices,
the port name does not show up, only the device description does.
The port name is not used very often around interfacing
with the audio backend as opposed to the device description, which is
typically used in various places around third-party apps.

This reverts commit 09358d26982431702c3652e8f07ebff7166a1181.