Display CPU usage as a bar graph using Unicode characters
Extend configuration keywords "%usage" with "%usagebar" and
"%cpu<n>" with "%cpubar<n>", which display the corresponding
statistics as a bar, one character wide, instead of a percentage.
These keywords can be combined in a single configuration.

The graph is relatively coarse, with a resolution of 11 percentage
points per bar increment, but it compresses the width of the
representation from (typically) four characters to one. This is
particularly useful on machines with larger numbers of cores.