Add memory usage for OpenBSD
Draw from sysctl(2) to populate memory fields for OpenBSD.  Unlike on
Linux, there's no shared or buffer memory, so the available memory is
simply free and cached.  This makes the memory calculation devices
return the same amount.

Update the manpage to reflect this new information, and also remove a
line specific to Linux (regarding returned amounts being in KiB) that
isn't relevant.